Counting islands
We need to count islands and make sure that they are connected horizontally or vertically. 
First of all, we need to create the function for calculating the number of islands and immediately check if matrix is not empty. Then we creating a functoin to calculate the area of the ocean and do the checking. Hence we calculate neighbors with four directions.  After all, we create a matrix with random numbers 0,1 and print a solution.